requirements.txt
# Requires Python 3.10

streamlit==1.45.1
tensorflow==2.19.0
numpy==2.1.3
Pillow==10.3.0
